remove router group
